add event tracking system, subject to change
