+++ sandbox_test/src/main.rs
@@ -1,8 +1,9 @@
 use sandbox::{
     bases::Baserunners,
     entities::{World},
+    events::{Event, Events},
     rng::Rng,
-    sim::{Event, Sim},
+    sim::Sim,
     mods::{Mod, ModLifetime},
     Game, GameTeam,
 };
@@ -10,18 +11,18 @@ use sandbox::{
 fn main() {
     //let mut rng = Rng::new(69, 420);
     //let mut rng = Rng::new(2200200200200200200, 1234567890987654321);
-    //let mut rng = Rng::new(3141592653589793238, 2718281828459045235);
-    let mut rng = Rng::new(37, 396396396396);
+    let mut rng = Rng::new(3141592653589793238, 2718281828459045235);
+    //let mut rng = Rng::new(37, 396396396396);
     //let mut rng = Rng::new(1923746321473263448, 2938897239474837483);
 
     let mut world = World::new();
     let team_a = world.gen_team(&mut rng, "Team A".to_string(), "A".to_string());
     let team_b = world.gen_team(&mut rng, "Team B".to_string(), "B".to_string());
 
-    world.player_mut(world.team(team_a).rotation[0]).mods.add(Mod::DebtU, ModLifetime::Permanent);
+    //world.player_mut(world.team(team_a).lineup[0]).mods.add(Mod::Shelled, ModLifetime::Permanent);
 
     let mut game = Game {
-        weather: sandbox::Weather::Eclipse,
+        weather: sandbox::Weather::Salmon,
         top: true,
         inning: 1,
         home_team: GameTeam {
@@ -41,11 +42,10 @@ fn main() {
         balls: 0,
         strikes: 0,
         outs: 0,
-        events_inning: 0,
         polarity: false,
         scoring_plays_inning: 0,
         salmon_resets_inning: 0,
-        last_salmon_inning: 0,
+        events: Events::new(),
         runners: Baserunners::new(),
         linescore_home: vec![0.0],
         linescore_away: vec![0.0],
@@ -54,7 +54,6 @@ fn main() {
     loop {
         let mut sim = Sim::new(&mut world, &mut rng);
         let evt = sim.next(&game);
-        game.events_inning += 1;
 
         // keeping sim outside the loop means it borrows world and we can't pass it as mut here, which might be fine...?
         evt.apply(&mut game, &mut world);
@@ -62,7 +61,7 @@ fn main() {
         if let Event::GameOver = evt {
             println!("game over!");
             break;
-        }
+        } 
 
         let base = format!(
             "[{}|{}|{}]",
